+++ src/storage/fts/ob_fts_parser_helper.cpp
@@ -19,6 +19,7 @@
 #define USING_LOG_PREFIX STORAGE_FTS
 
 #include "storage/fts/ob_fts_parser_helper.h"
+#include <cinttypes>
 
 #include "common/json_type/ob_json_tree.h"
 #include "share/ob_force_print_log.h"
@@ -116,7 +117,8 @@ int ObFTParser::serialize_to_str(char *buf, const int64_t buf_len)
   } else if (OB_UNLIKELY(!is_valid())) {
     ret = OB_NOT_SUPPORTED;
     LOG_WARN("invalid fulltext parser doesn't support to serialize_to_str", K(ret), KPC(this));
-  } else if (OB_FAIL(common::databuff_printf(buf, buf_len, pos, "%.*s.%ld", parser_name_.len(), parser_name_.str(),
+  } else if (OB_FAIL(common::databuff_printf(buf, buf_len, pos, "%.*s.%" PRId64,
+          static_cast<int>(parser_name_.len()), parser_name_.str(),
           parser_version_))) {
   }
   return ret;
